correct migrations id type to integer

+++ db/migrate/20210629114124_change_types_issues.rb
@@ -1,6 +1,6 @@
 class ChangeTypesIssues < ActiveRecord::Migration[6.1]
   def change
-    change_column :issues, :category_id, :bigint
-    change_column :issues, :user_id, :bigint
+    change_column :issues, :category_id, :integer
+    change_column :issues, :user_id, :integer
   end
 end
