+++ db/migrate/20210629114500_change_types_rooms_services.rb
@@ -1,6 +1,6 @@
 class ChangeTypesRoomsServices < ActiveRecord::Migration[6.1]
   def change
-    change_column :rooms_services, :room_id, :bigint
-    change_column :rooms_services, :service_id, :bigint
+    change_column :rooms_services, :room_id, :integer
+    change_column :rooms_services, :service_id, :integer
   end
 end
